Build Dashboard parent widget
